ShadowTrace_backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
python-dotenv
pymongo
elasticsearch
rapidfuzz
networkx
weasyprint
Jinja2
python-whois
shodan
python-dotenv
pymongo
fastapi
uvicorn
